requirements-api.txt
# Core Banking System - Mobile Banking API Requirements

Flask==2.2.3
Flask-Cors==3.0.10
PyJWT==2.6.0
cryptography==40.0.1
jsonschema==4.17.3
marshmallow==3.19.0
mysql-connector-python==8.0.32
python-dotenv==1.0.0
requests==2.28.2
Werkzeug==2.2.3
Flask-Limiter==3.3.0
